refactor: moved Image and Pluralizer to Components folder
